DPC-276: Unify Hibernate entity conversion utils. (#504)
* DPC-276: Initial commit of FHIRConverter engine.

* DPC-276: OrganizationEntity is now converting from FHIR using the new converter framework.

* DPC-276: Added very simple organization conversion test.

We now 'hopefully' have an abstract test framework for verifying conversion utilities.

* DPC-276: Finished migrating entity converts to the new unified framework.

There are still some classes that we can't fully convert because they require more context at runtime.

* DPC-276: Removed all the old converters and function signatures

* DPC-276: Some more refactoring and code removal.

Also, error prone fixes.

* DPC-276: Improve unit test coverage and add exception handling for converters.

Internally, everything is a mess, but it definitely has a cleaner interface than before.

* Code Climate cleanups.

* DPC-276: Add javadoc

* DPC-276: Address code review comments.

* DPC-276: Remove and comment unused unit tests

Removed a duplicated entity conversion test and added a comment as to why the other one isn't enabled yet. We'll re-add it as part of DPC-935, but for right now, it's a placeholder.