DPC-453: Submit attribution attestation (#516)
* DPC-453: Initial commit of Roster Attestation

Wired in the recording of the attribution attestation into the Group resource.

* DPC-453: Moved patient and practitioner tests to using the submit endpoint

Rather than doing things transactionally (which is really slow) we can do them in a single batch. The smoke tests already validate the individual requests work in a reasonable manner, so we don't need to double test it.

* DPC-453: Update logging and add provenance to all Group write methods

Logging more information and doing it for more methods.

Updated the FHIRSubmissionTest to include a mock Provenance injector.

* DPC-453: Initial commit for DPC Attestation FHIR profile

* DPC-453: Updated group resource to validate Attestation profile

* DPC-453: Updated IG with the attestation profile.

* DPC-453: Fix failing unit tests.

Updated the unit tests to account for the newly added FHIR profile.

* DPC-453: Update Postman tests to support provenance attestations

Also disabled the default schema and schematron validation.
It would be great to have this enabled, but we can't do so because we remove the requirement for a Target resource in our Provenance resources. We probably need a better workaround, but ok for now.

* DPC-453: Fix Target count in profile

It is possible to have a target in the provenance resource, so we shouldn't rule that out.

* DPC-453: Added provenance header to Group resource Swagger docs

* DPC-453: Added provenance information to documentation.

* DPC-453: Fix parameter ordering in attestation logging.

* DPC-453: Added unit test to verify log output.

Turns out, the logging was completely broken, but fortunately, we have good engineers that catch my mistakes!

We're logging everything for each add operation, and when possible, we get the roster ID as well.

* DPC-453: Remove nullable annotation.

Having the annotation throws off Sonar's null check logic, because it thinks the variable itself could be null, rather than just the parameter to the method.

* DPC-453: Fix error prone