DPC-248: Enable distributed tracing (#556)
* DPC-248: Enable distributed tracing

Update the newrelic.yml file to enable distributed tracing.

* DPC-248: Disable cross app tracing and enable trace logging

Try to make things more verbose, as well as disable cross application tracing

* DPC-248: Debug logging only for http headers

* DPC-248: Temporarily disable circuit breakers.

* DPC-248: Enable distributed tracing for website.

Re-enable circuit breaker, but increase the GC limit to so that it doesn't die on us.

* DPC-248: Add manual trace transaction to aggregation service

Updated the aggregation engine to manually trace aggregation jobs.

Not sure which level of granularity we need right now, but this should be a good start.

* DPC-248: Remove debug logging of HTTP headers

In order to avoid leaking PII into the logs, I'm removing the debug header output. We may want to re-enable this in the future, in order to get some additional visibility into the intra-DPC service traffic.