Rename MBI identifier to BENE_ID (#557)
* Rename MBI identifier to BENE_ID.

In preparation for our migration to MBIs from BB bene_id, we need to rename the identifier enum to avoid conflicts in the future.

This is a simple IntelliJ refactoring.

* Update identifier system

Updated the identifier system in the ConsentResourceTest parameterized values.
We'll need to switch this back when we move fully to MBIs.

* DPC-942: Update organization in API from web app (#555)

* Update organization from web app to API

Use FHIR client to update registered organizations in API

Also some cleanup from manual testing; remove delay in org creation

* Add fhir endpoint updating

* Add payloadType to endpoint update request

* Add Endpoint.payloadType to Organization$submit

This field is required by FHIR but currently our API has a bug that
skips validation of this field when creating an Endpoint. This commit
adds the field and updates the test stubs.

See:
https://hl7.org/fhir/STU3/endpoint-example.json.html
https://dpc.cms.gov/ig/StructureDefinition-dpc-profile-endpoint-definitions.html

* Cleanup

* Refactor for smaller APIClient class

Pull out FHIR resource building into a separate class.

One other cleanup fix in attribute method in organization_registrar

More cleanup from codeclimate recs

* Refactor APIClient to separate org and endpoint

Org and endpint update requests should be separate methods because
they are separate interactions with the API.

* Cleanup ProfileTests

Removed some unused variables and added todo to update the MBI's, once we're ready for it.

* Revert patient profile changes.

This should not have been committed yet. It will come in a later PR.

* Revert patient profile test.

Co-authored-by: Shelby Switzer <shelby.switzer@cms.hhs.gov>