Update tests to account for javascript followup
We use javascript to dynamically hide/show fields in these forms, but
in order to replicate that functionality in tests, we would need to
set up selenium and capybara properly. That is out of scope for this
PR, so for now we will mark these fields as not visible (though in
actuality they are triggered to be visible with JavaScript).