DEPLOY_ENV organization conditional added
If DEPLOY_ENV == prod-sbx the app will generate a vendor_id for a vendor org or a provider_id if no NPI provided.

If DEPLOY_ENV is not prod-sbx organizations will require an NPI to be enabled in the API.

If no NPI provided, the app will generate the error "NPI required".