DPC-902: Fix return value when adding too many providers (#1123)
* fix: improve test to simulate bulk upload with no existing providers
When onboarding a vendor, we hit an indexOutOfBounds error because they
were doing their initial upload of providers with a number greater than
the limit. This test was not simulating that potential issue correctly
because it used the same provider both times, meaning that there were
existing providers with that NPI already, thus not causing this error.

* fix: return the entity when over limit instead of searching

* fix: remove unused imports