chore: remove unused swagger annotations from attribution
we use swagger annotations in api to autogenerate docs
for public endpoints. however, we do not want to generate
docs for the underlying attribution methods.
For this reason, we need to remove the unecessary annotations
as well as the test in attribution that was checking for them.
This commit also includes some whitespace cleanup.
DPC-1095