use `DecryptedDir` EV for certificate locations
Because the certificates are accessed from different levels of subdirectories, it is no longer possible to use relative links to the files in the config. Instead, we now construct the location using the `DecryptedDir` and the file name from the config. Required updating expected error messages in the client tests.