Gosec error fix (#1808)
Ignoring this line as research has indicated this is not a problem, connections are recommended to be closed, and this code is not used or maintained at this time as well. V2 was never supported or worked on by the Fearless team. Moving V1 to a production ready state was priority per CMS.