Me dpc 3494 part 2 upgrade mockito (#1911)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3494

## 🛠 Changes

Updated Mockito from 3.4.6 to 5.4.0.
Removed mocking of Method class from FHIRRequestFeatureTest.
Updated BaseResourceUnitTest to use new MockedStatic and
MockedConstruction.

## ℹ️ Context for reviewers

We were running an old version of Mockito that didn't support mocking
static or final methods, or constructors, and this was preventing us
from pushing code coverage as far as we otherwise could. The new version
of Mockito was also causing unit tests to fail randomly and erratically.

The new version was causing tests to fail because an old test in
`FHIRRequestFeatureTest` was mocking the `Method` class from
`java.lang.reflect`. Mockito depends on this class, along with a few
others from that package, and mocking them is essentially rewriting
Mockito's plumbing while it's running. This leads to any tests running
afterwards potentially failing in random ways. See issue
[here](https://github.com/mockito/mockito/issues/2026#issuecomment-741663017)
for a better explanation.

## ✅ Acceptance Validation

We're running the new version of Mockito, we're using some of its new
features, and all unit tests are passing.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.