Me dpc-3512 add logging attributes (#1925)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3512

## 🛠 Changes

Adds the environment, application and version to all of our logs.

## ℹ️ Context for reviewers

This PR relies on CMSGov/dpc-ops#585 to set the environment vars.  

## ✅ Acceptance Validation

Branch was deployed to dev and the new values are in the logs.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.