Me dpc 3576 fix logging (#1940)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3576

## 🛠 Changes

1. Updated TokenResource so that when `/auth` is called with the same
JWT in succession (a JWT replay attack), the error logged will include
the correct organization id. The org id is pulled from the macaroon sent
as part of the JWT.
2. Updated default logging so that when no APPLICATION env variable is
set it uses real names by default. I.E., when running locally in Docker,
the service tag in the logs will say "dpc-api" instead of
"unknown_service".
3. Added make targets `start-dpc-debug` and `start-app-debug` that
function just like `start-dpc` and `start-app`, but they first recompile
dpc with debug extensions and start each service with the debugging
agent enabled. This lets you connect a remote debugger to them and made
working on this ticket much, much easier.

## ℹ️ Context for reviewers

In the case of a JWT being replayed, the error logged included a random
org id instead of the real one contained in the JWT.

## ✅ Acceptance Validation

Ran locally and observed the new logs.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.