[DPC-3599] Opt-out Import Database Migrations (#1941)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3599

## 🛠 Changes

This PR does the following:
1. A database migration for `dpc_consent` db. It creates a table called
`opt_out_file`, consisting of metadata for each opt out file we import,
as well as a foreign key `opt_out_file_id` on the `consent` table. For
every beneficiary who chooses to opt out, we'll be storing a record in
the `consent` table. The Many-to-One relationship here allows us to tie
each record to an opt out file and timestamp, for
visibility/transparency.
2. Creates an `OptOutFileEntity` for the ORM, so that our Java code can
interact with it.
3. Updates the `ConsentEntity` to reflect the Many-to-One relationship.

## ✅ Acceptance Validation

- Successfully ran the migration on my local environment without any
issues.
- Successfully compiled `dpc-common` and `dpc-consent`.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.

---------

Co-authored-by: Ashley Weaver <134093673+ashley-weaver@users.noreply.github.com>