[dpc-3630] Jd me/dpc 3630 add dpc portal (#1954)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3630

## 🛠 Changes

Setup framework for the new dpc-portal app.
Updated docker-compose.portals to start new app.
Added portal DB and connected it to app.
Updated make targets to start, stop and test new app.
Updated Git workflow to run doc-portal tests.

## ℹ️ Context for reviewers

We're building a new doc-portal that will replace dpc-web and dpc-admin,
and this is the first step.
We used the same Gemfile as is used in the other rails services, as they
now use rails 7. We left out the ActionCable gem, as it had been removed
because of a security issue in rails 5, but we could simplify the
Gemfile (and thus future upgrades) by putting it back in. We also left
in other gems, like Devise (authentication) and fihr for their potential
future use, but maybe they should be removed?

## ✅ Acceptance Validation

The portal runs locally, serves up a "Hello World" page, and all tests
pass locally and on GitHub.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.

---------

Co-authored-by: Jeffrey Dettmann <jeffreydettmann@navapbc.com>
Co-authored-by: Kevin Yeh <kevinyeah@utexas.edu>