DPC-3636: Extract api client as shared engine for dpc-admin, dpc-web, and dpc-portal (#1961)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3636

## 🛠 Changes

Api Client (with helpers FhirResourceBuilder and
OrganizationSubmitSerializer) abstracted from dpc-web and dpc-admin into
a separate engine to be included in both projects, as well as
dpc-portal.

Api Client renamed to DpcClient to make more clear what it is,
especially given the new PECOS client on the horizon.

## ℹ️ Context for reviewers

The bulk of the code for this review is generated by Rails as part of
the engine creation process. The critical pieces of code are:
- engines/api_client/app/services/dpc_client.rb
- engines/api_client/app/services/fhir_resource_builder.rb
- engines/api_client/spec/services/dpc_client_spec.rb
- {dpc-admin,dpc-web}/app/models/registered_organization.rb

~~We don't have a gem server, so we are loading the gem via git. We will
need to add another pull request for the api_client alone in order to
merge it into master, so we can update the Gemfiles to point to the
master branch and have everything still work.~~ (Now loading from path,
so this is not longer true.)

## ✅ Acceptance Validation
First, I set the dpc app running. then I got a new golden macaroon.

In the rails console for both dpc-web and dpc-admin, I did the
following:

- Updated the API_METADATA_URL and GOLDEN_MACAROON to match the running
dpc app
- Created an Organization
- Created a RegisteredOrganization linked to the above Organization and
the default fhir endpoint.
- Validated that the organization was now present in the dpc app
database
- Validated I received the appropriate information from the
DpcClient#get_organization with the new RegisteredOrganization#api_id
- Updated the Organization and verified that the changes where in the
dpc app database
- Updated the RegisteredOrganization (changed "enabled" back and forth)
and confirmed that both both update_organization and update_endpoint
where called
- Built a new ClientTokenManager with the new RegisteredOrganization
- Created two new client_tokens via the ClientTokenManager
- Retrieved the new client_tokens via the ClientTokenManager
- Deleted one of the new client_tokens via the ClientTokenManager
- Built a new PublicKeyManager
- Followed the instructions for creating a public pem and signature file
from dpc-web/app/views/public_views/new.html.erb
- Created a public_key with these parameters via the PublicKeyManager
- Retrieved the public_key via the PublicKeyManager
- Deleted the public_key via the PublicKeyManager (dpc-web only, as this
is not available in dpc-admin)

## 🔒 Security Implications

- [x] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.