DPC-3695 Resolve Smoke Test Failures (#1976)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3695

## 🛠 Changes

- SmokeTest getOrg was updated to accept 500 as well as 404 errors as
assurance Organization did not exist
- SmokeTest makes sure all tests actually pass

## ℹ️ Context for reviewers

The [DPC Smoke Test
](https://management.dpc.cms.gov/job/DPC%20-%20Smoke%20Test/) kept
failing when run against prod-sbx. As part of the setup, SmokeTest
checks to see if an organization exists and deletes it if it doesn't.
The usually cause of failure was due to HAPIExceptionHandler in dpc-api
sometimes failed to consider
ca.uhn.fhir.rest.server.exceptions.ResourceNotFoundException to be a
FHIRResource, and would return a 500 instead of 404, which would cause
the test to fail.

When debugging, I noticed that the tests sometimes would pass even if
something untoward happened, such as the failure to upload the
publicKey. I therefore updated the tests to fail unless everything that
was supposed to happen did happen.