DPC-3740: Set up frontend tooling and CSS pipeline (#1979)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3740

## 🛠 Changes

### Tooling Changes

- Remove webpacker, which was deprecated and replaced by other tooling
in Rails 7
- Replace webpacker SASS --> CSS pipeline with basic sass compilation
- Add ViewComponent, USWDS and Lookbook integration/setup
- Update docker setup to enable live-reload (updating application
without rebuilding docker container)
- Remove unused NPM dependencies

### Minor adjustments

- Update application layout to include grid container setup from USWDS
(i.e. add global margins for pages)
- Move devise users endpoints to live within the portal scope, to ensure
they are routed properly in deployed environments

## ℹ️ Context for reviewers

We talked through some frontend work in refinement, and I realized that
we still had more cleanup and groundwork to do than I expected. In
particular:

- Webpacker was deprecated in Rails 7 and replaced with a set of
different options (import maps, jsbundling-rails + cssbundling-rails,
shakapacker.)
- For now I've only migrated the SASS --> CSS pipeline and not the JS
pipeline (since we aren't leveraging any JS libraries at the moment that
we need to compile.)
- We didn't have a pattern set up for Viewcomponent + USWDS. I set up an
example Alert component to demonstrate the folder structure with tests
and a Lookbook preview.

### Asides

- I had attempted to import the uswds_components library but it was not
available on rubygems — opened an issue with them to see if they can fix
that.
- I also noted that `minitest` is the test library of choice by Rails
now, but I have stuck with an RSpec test since it is already set up and
configured for code coverage. We may wish to move to minitest in the
near future.

## ✅ Acceptance Validation

Tested development locally. I have a slightly out-of-date gif below of
the live reload + Lookbook development process. To replicate...

1. Check out this branch
2. Run `make portal` and `make start-portals`
3. Visit localhost:3001/portal/lookbook


![live-reload](https://github.com/CMSgov/dpc-app/assets/2308368/14d03766-6aad-4305-9cdf-d39f337653b8)

## 🔒 Security Implications

- [x] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

**NOTE** 

@StewGoin -- This PR adds mainly development-only dependencies, _except_
for ViewComponent which was previously discussed and approved in
https://jira.cms.gov/browse/DPC-3631.