[DPC-3683] Initial Client Code for CPI API Gateway (#1983)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3683

## 🛠 Changes

This PR establishes a client class that, upon instantiation,
authenticates with CMS IDM in the Val environment. As API calls are
implemented, this class will continue to manage the access token.

## ℹ️ Context for reviewers

In order to implement an RSpec test for this using VCR, as discussed,
the Docker container I'm using needs to be able to connect to the VPN
via Zscaler, so that it can make the appropriate API calls. Given that I
haven't been able to make this happen, I've opted to exclude unit tests
from this PR. Tests will be added in DPC-3684.


## ✅ Acceptance Validation

(How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable.)

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.