DPC-3740: Install lookbook in deployed environments (#1986)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3740

## 🛠 Changes

- Move lookbook dependency from :development group to :production group
in Gemfile

## ℹ️ Context for reviewers

In #1979 we installed lookbook and configured it to deploy in the dev
environment. However, the lookbook dependency is only installed in the
:development RAILS_ENV group. It needs to be installed in the
:production group, which is the RAILS_ENV used in all deployed
environments.

In the future we should make sure this dependency is not installed in
sandbox/prod, but until we get to production launch readiness we can
keep this in the production group.

## ✅ Acceptance Validation


## 🔒 Security Implications

- [x] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

cc. @StewGoin