DPC-3691 Base for Manage organizations (#1992)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3691

## 🛠 Changes

- Organizations Controller
- Organization model
- PublicKeyManager
- ClientTokenManager
- specs for above


## ℹ️ Context for reviewers

This is not for the DPC-3691 functionality. This is just a base so
fellow developers will not have difficulties merging

## ✅ Acceptance Validation

(How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable.)

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.