[DPC-3761] Me/dpc 3761 add ip address table (#2003)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3761

## 🛠 Changes

Added ip_address table to the `dpc-auth` DB, along with `IpAddressDAO`
and `IpAddressEntity` classes in dpc-api.

Also added an AbstractDAOTest class to allow us to easily get code
coverage on our DAO objects.

## ℹ️ Context for reviewers

If you attempt to run this locally, you'll need to first delete your
`pgdata` volume in Docker and then run `make api` and `make start-api`.
This will remove your old version of the dpc-auth DB and rebuild
dpc-api, and then start dpc-api, which will run the new DB migration on
dpc-auth. At that point you should be able to connect to the DB and see
the new table.

`AbstractDAOTest` is a modified version of the `DAOTestExtension`
already provided by DropWizard, but instead of using an in memory H2
database it spins up PostgreSql in a test container. This allows us to
run tests against the same DB we use in production and avoid some
problems with H2 like not supporting the inet type. It should make it
pretty easy to go back and add code coverage to all of our DAO classes.

Side note: I originally wanted `AbstractDAOTest` to live in test/java
instead of src/main/java in dpc-common since it's only going to be used
for testing, but every time I tried to move it there and make it
available to the other sub-modules the .pom files started getting messy.
If anyone wants to take a crack at it be my guest! If not, maybe a
future improvement?

## ✅ Acceptance Validation

All tests are passing and we have 100% code coverage on the new DAO and
entity.

## 🔒 Security Implications

- [x] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.