Me/dpc 3860 hapi exception fix (#2078)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3860

## 🛠 Changes

- Updated `HAPIExceptionHandler` so that all exceptions it handles
return the correct HTTP status.
- Bumps Jib so that we can use the newest version of Docker without
getting the LayerSources error when attempting to build dpc.
- Fixed broken CLI tests that won't run locally, but were running in
GHA.

## ℹ️ Context for reviewers

The previous version of this code used an injected `ResourceInfo` object
to determine if the end point from which the exception was thrown was a
FHIR end point or not. If it was a FHIR end point the correct http
status was returned, and if it wasn't another exception was thrown and a
500 was returned. The problem is that at least some of the time, and
only in our persisted environments, Jersey failed to fill out the
`ResourceInfo` object. That lead to this handler returning 500's when it
shouldn't have.

In the new version, we don't care if the end point the exception was
thrown from was a FHIR end point or not. In either case, we return the
correct http status to the caller and fill out the response body with an
`OperationOutcome`.

Also, this handler only works on `BaseServerResponseExceptions`, which
are defined in HAPI FHIR. The only way to get a
`BaseServerResponseException` outside of a FHIR end point is to have a
non-FHIR end point make a FHIR call with the HAPI client and for it to
throw an exception. I'm not sure that's possible in our code, so
worrying about whether an exception comes from a FHIR end point or not
is probably moot.

## ✅ Acceptance Validation

Code updated and automated and manual tests are passing.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.