Automatically generate golden macaroon on local dev startup (#2014)
## 🎫 Ticket

N/A

## 🛠 Changes

- Update docker entrypoint scripts to generate golden macaroon on
startup

## ℹ️ Context for reviewers

This removes a manual step required to call the API and generate golden
macaroons. If the API is running when a frontend application starts, we
will automatically generate a new macaroon If the API is not running, we
will continue as normal and log a message noting that there is no
macaroon.

## ✅ Acceptance Validation

Ran `make portal && make start-portal` with and without a golden
macaroon set in the docker-compose file and observed both behaviors.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.