[DPC-3968] Use correct filename prefix for prod and lower environments (#2084)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3968

## 🛠 Changes

This updates the functions that handle filename generation and parsing
for both lambdas. Specifically, in production, we expect to receive and
generate files with a `P` prefix; for all lower environments, the prefix
should be `T`.

## ℹ️ Context for reviewers

(Background context, more in-depth details of the implementation, and
anything else you'd like to call out or ask reviewers.)

## ✅ Acceptance Validation

(How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable.)

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.